requirements.txt
asgiref==3.5.0
celery==5.1.2
Django==4.0.3
django-enumchoicefield==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
restframework-serializer-permissions==0.0.2
sqlparse==0.4.2
redis==4.2.0
python-dotenv==0.20.0
